+++ src/QueryCat.Cli/Commands/BaseQueryCommand.cs
@@ -2,6 +2,10 @@ using System.CommandLine;
 using QueryCat.Backend.Core;
 using QueryCat.Backend.Core.Execution;
 using QueryCat.Backend.Core.Types;
+using QueryCat.Backend.Core.Utils;
+using QueryCat.Backend.Execution;
+using QueryCat.Backend.ThriftPlugins;
+using QueryCat.Cli.Infrastructure;
 
 namespace QueryCat.Cli.Commands;
 
@@ -38,12 +42,37 @@ internal abstract class BaseQueryCommand : BaseCommand
         {
             foreach (var file in files)
             {
-                executionThread.Run(File.ReadAllText(file), cancellationToken: cancellationToken);
+                RunWithPluginsInstall(executionThread, File.ReadAllText(file), cancellationToken: cancellationToken);
             }
         }
         else
         {
+            RunWithPluginsInstall(executionThread, query, cancellationToken: cancellationToken);
+        }
+    }
+
+    private void RunWithPluginsInstall(IExecutionThread executionThread, string query, CancellationToken cancellationToken)
+    {
+        try
+        {
             executionThread.Run(query, cancellationToken: cancellationToken);
+        }
+        catch (ProxyNotFoundException)
+        {
+            Console.WriteLine(Resources.Messages.PluginProxyWantToInstall);
+            var key = Console.ReadKey();
+            if (key.Key == ConsoleKey.Y)
+            {
+                var applicationDirectory = ExecutionThread.GetApplicationDirectory(ensureExists: true);
+                var pluginsProxyLocalFile = Path.Combine(applicationDirectory,
+                    ThriftPluginsLoader.GetProxyFileName(includeCurrentVersion: true));
+                AsyncUtils.RunSync(ct =>
+                {
+                    var downloader = new PluginProxyDownloader(ThriftPluginsLoader.GetProxyFileName());
+                    return downloader.DownloadAsync(pluginsProxyLocalFile, ct);
+                });
+                executionThread.Run(query, cancellationToken: cancellationToken);
+            }
         }
     }